Created `NestedScrollableContent`
This let's us use any Composable in the scrollable gradient thing.
